# openapi-appengine.yaml
swagger: "2.0"
...
info:
  description: "Dantion Endpoints API."
  title: "Endpoints Dantion"
  version: "1.0.0"
host: "danger-detection.appspot.com"
...
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - "https"
  - "http"

# security:
#   - api_key: []
securityDefinitions:
  # api_key:
  #   type: "apiKey"
  #   name: "key"
  #   in: "query"
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
paths:
...
  "/users/{id}":
    get:
      tags:
        - "Users"
      security:
        - Bearer: []
      description: "Get User Using ID"
      operationId: "Userid"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: path
          description: "ID of user"
          required: true
          type: "string"
      responses:
        201:
          description: "success get detail User"
          schema:
            $ref: "#/definitions/UserSuccessResponse"
        400:
          description: "failed get detail User"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
...
definitions:
  StatusMessageResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      message:
        type: "string"
...
  User:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      address:
        type: "string"
      number:
        type: "string"
      parentNumber:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
        format: "password"
      role:
        type: "string"
        enum:
          - "admin"
          - "polisi"
          - "ambulan"
          - "damkar"
      photo:
        type: "string"
        format: "binary"
      createdAt:
        type: "string"
        format: "date-time"
      updatedAt:
        type: "string"
        format: "date-time"
...
  UserSuccessResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      message:
        type: "string"
      user:
        $ref: "#/definitions/User"
...